+++ AtlasAgent/src/atlas-agent.cpp
@@ -1,45 +1,24 @@
-#ifdef __linux__
-#include <sys/vfs.h>
-#endif
-#include <thirdparty/spectator-cpp/spectator/registry.h>
-#include <lib/collectors/amd_smi/gpumetrics.h>
-#include <lib/collectors/aws/src/aws.h>
-#include <lib/collectors/cgroup/src/cgroup.h>
-#include <lib/collectors/cpu_freq/src/cpu_freq.h>
-#include <lib/collectors/dcgm/src/dcgm_stats.h>
-#include <lib/collectors/disk/src/disk.h>
-#include <lib/collectors/ebs/src/ebs.h>
-#include <lib/collectors/ethtool/src/ethtool.h>
-#include <lib/collectors/nvml/src/gpumetrics.h>
-#include <lib/collectors/ntp/src/ntp.h>
-#include <lib/collectors/perf_metrics/src/perf_metrics.h>
-#include <lib/collectors/perfspect/src/perfspect.h>
-#include <lib/collectors/pressure_stall/src/pressure_stall.h>
-#include <lib/collectors/proc/src/proc.h>
-#include <lib/collectors/service_monitor/src/service_monitor.h>
+// Entry point for the Atlas agent. This file holds the pieces common to both
+// build flavors — option parsing, signal handling, and the helpers shared by
+// the collector loops. The per-flavor collection logic lives in
+// system-agent.cpp / titus-agent.cpp (only one is compiled per build, selected
+// by TITUS_SYSTEM_SERVICE); the shared declarations live in atlas-agent.h.
+
+#include "atlas-agent.h"
+
+#include <lib/http_client/src/http_client.h>
 #include <lib/util/src/util.h>
 
 #include "backward.hpp"
-#include <condition_variable>
+
+#include <cassert>
 #include <csignal>
-#include <fmt/chrono.h>
+#include <cstdio>
+#include <cstdlib>
+#include <exception>
 #include <getopt.h>
 #include <random>
-
-using atlasagent::GetLogger;
-using atlasagent::Logger;
-using atlasagent::Nvml;
-
-using Aws = atlasagent::Aws;
-using CGroup = atlasagent::CGroup;
-using CpuFreq = atlasagent::CpuFreq;
-using Disk = atlasagent::Disk;
-using Ethtool = atlasagent::Ethtool;
-using GpuMetrics = atlasagent::GpuMetrics<Nvml>;
-using Ntp = atlasagent::Ntp<>;
-using PerfMetrics = atlasagent::PerfMetrics;
-using PressureStall = atlasagent::PressureStall;
-using Proc = atlasagent::Proc;
+#include <utility>
 
 std::unique_ptr<GpuMetrics> init_gpu(Registry* registry, std::unique_ptr<Nvml> lib)
 {
@@ -57,83 +36,6 @@ std::unique_ptr<GpuMetrics> init_gpu(Registry* registry, std::unique_ptr<Nvml> l
     }
     return {};
 }
-
-#if defined(TITUS_SYSTEM_SERVICE)
-static void gather_peak_titus_metrics(CGroup* cGroup, const bool fiveSecondMetricsEnabled, const bool sixtySecondMetricsEnabled)
-{ 
-    cGroup->CpuStats(fiveSecondMetricsEnabled, sixtySecondMetricsEnabled);
-}
-
-static void gather_slow_titus_metrics(CGroup* cGroup, Proc* proc, Disk* disk, Aws* aws)
-{
-    aws->update_stats();
-    cGroup->MemoryStatsV2();
-    cGroup->MemoryStatsStdV2();
-    cGroup->NetworkStats();
-    disk->titus_disk_stats();
-    proc->netstat_stats();
-    proc->network_stats();
-    proc->process_stats();
-    proc->snmp_stats();
-    proc->uptime_stats();
-}
-#else
-static void gather_peak_system_metrics(Proc* proc, const bool fiveSecondMetricsEnabled, const bool sixtySecondMetricsEnabled)
-{
-    proc->CpuStats(fiveSecondMetricsEnabled, sixtySecondMetricsEnabled);
-}
-
-static void gather_scaling_metrics(CpuFreq* cpufreq) { cpufreq->Stats(); }
-
-static void gather_slow_system_metrics(Proc* proc, Disk* disk, Ethtool* ethtool, Ntp* ntp, PressureStall* pressureStall,
-                                       Aws* aws)
-{
-    aws->update_stats();
-    disk->disk_stats();
-    ethtool->update_stats();
-    ntp->update_stats();
-    pressureStall->update_stats();    
-    proc->arp_stats();
-    proc->loadavg_stats();
-    proc->memory_stats();
-    proc->netstat_stats();
-    proc->network_stats();
-    proc->process_stats();
-    proc->snmp_stats();
-    proc->socket_stats();
-    proc->uptime_stats();
-    proc->vmstats();
-}
-#endif
-
-struct terminator
-{
-    terminator() noexcept = default;
-
-    // returns false if killed:
-    template <class R, class P>
-    bool wait_for(std::chrono::duration<R, P> const& time)
-    {
-        if (time.count() <= 0)
-        {
-            Logger()->warn("waiting for zero ticks!");
-            return true;
-        }
-        std::unique_lock<std::mutex> lock(m);
-        return !cv.wait_for(lock, time, [&] { return terminate; });
-    }
-    void kill()
-    {
-        std::unique_lock<std::mutex> lock(m);
-        terminate = true;
-        cv.notify_all();
-    }
-
-   private:
-    std::condition_variable cv;
-    std::mutex m;
-    bool terminate = false;
-};
 
 terminator runner;
 
@@ -198,273 +100,6 @@ long initial_polling_delay()
         return 0;
     }
 }
-
-#if defined(TITUS_SYSTEM_SERVICE)
-void collect_titus_metrics(Registry* registry, std::unique_ptr<atlasagent::Nvml> nvidia_lib,
-                           const std::unordered_map<std::string, std::string>& net_tags,
-                           const int& max_monitored_services)
-{
-    using std::chrono::duration_cast;
-    using std::chrono::milliseconds;
-    using std::chrono::seconds;
-    using std::chrono::system_clock;
-
-    Aws aws{registry};
-    CGroup cGroup{registry};
-    Disk disk{registry, ""};
-    PerfMetrics perf_metrics{registry, ""};
-    Proc proc{registry, std::move(net_tags)};
-
-    auto gpu = init_gpu(registry, std::move(nvidia_lib));
-
-    // TODO: DCGM & ServiceMonitor have Dynamic metric collection. During each iteration we have to
-    // check if these optionals have a set value. lets improve how we handle this
-    std::optional<ServiceMonitor> serviceMetrics{};
-    std::optional<std::vector<std::regex> > serviceConfig{
-        parse_service_monitor_config_directory(ServiceMonitorConstants::ConfigPath)};
-    if (serviceConfig.has_value())
-    {
-        serviceMetrics.emplace(registry, serviceConfig.value(), max_monitored_services);
-    }
-    else
-    {
-        Logger()->info("Service Monitoring is disabled.");
-    }
-
-    // initial polling delay, to prevent publishing too close to a minute boundary
-    auto delay = initial_polling_delay();
-    Logger()->info("Initial polling delay is {}s", delay);
-    if (delay > 0)
-    {
-        runner.wait_for(seconds(delay));
-    }
-
-    // the first call to this gather function takes ~100ms, so it must be
-    // done before we start calculating times to wait for peak metrics
-    gather_slow_titus_metrics(&cGroup, &proc, &disk, &aws);
-    Logger()->info("Published slow Titus metrics (first iteration)");
-
-    auto now = system_clock::now();
-    auto next_run = now;
-    auto next_sixty_second_run = now + seconds(60);
-    auto next_five_second_run = now + seconds(5);
-    std::chrono::nanoseconds time_to_sleep;
-
-    do
-    {
-        auto start = system_clock::now();
-        bool fiveSecondMetricsEnabled = (start >= next_five_second_run);
-        bool sixtySecondMetricsEnabled = (start >= next_sixty_second_run);
-
-        // 1 second, 5 second, and 60 second CPU metrics are gathered here because they read from
-        // the same /proc/stat file
-        gather_peak_titus_metrics(&cGroup, fiveSecondMetricsEnabled, sixtySecondMetricsEnabled);
-
-        // If its time to gather 5 second metrics, update the next run time
-        // Currently we only have CPU metrics that run every 5 seconds, but if we add more in the future
-        // we can gather them here
-        if (fiveSecondMetricsEnabled == true)
-        {
-            cGroup.IOStats();
-            next_five_second_run += seconds(5);
-        }
-
-        // If its time to gather 60 second metrics, gather the metrics and update the next run time
-        if (sixtySecondMetricsEnabled == true)
-        {
-            gather_slow_titus_metrics(&cGroup, &proc, &disk, &aws);
-            perf_metrics.collect();
-            if (gpu)
-            {
-                gpu->gpu_metrics();
-            }
-            if (serviceMetrics.has_value() && serviceMetrics.value().gather_metrics() == false)
-            {
-                Logger()->error("Failed to gather Service metrics");
-            }
-            auto elapsed = duration_cast<milliseconds>(system_clock::now() - start);
-            Logger()->info("Published Titus metrics (delay={})", elapsed);
-            next_sixty_second_run += seconds(60);
-        }
-
-        next_run += seconds(1);
-        time_to_sleep = next_run - system_clock::now();
-    } while (runner.wait_for(time_to_sleep));
-}
-#else
-void collect_system_metrics(Registry* registry, std::unique_ptr<atlasagent::Nvml> nvidia_lib,
-                            const std::unordered_map<std::string, std::string>& net_tags,
-                            const int& max_monitored_services)
-{
-    using std::chrono::duration_cast;
-    using std::chrono::milliseconds;
-    using std::chrono::seconds;
-    using std::chrono::system_clock;
-
-    Aws aws{registry};
-    CpuFreq cpufreq{registry};
-    Disk disk{registry, ""};
-    Ethtool ethtool{registry, net_tags};
-    Ntp ntp{registry};
-    PerfMetrics perf_metrics{registry, ""};
-    PressureStall pressureStall{registry};
-    Proc proc{registry, net_tags};
-
-    auto gpu = init_gpu(registry, std::move(nvidia_lib));
-
-    std::optional<GpuMetricsDCGM> gpuDCGM{std::nullopt};
-    if (atlasagent::is_file_present(DCGMConstants::dcgmiPath))
-    {
-        gpuDCGM.emplace(registry);
-    }
-
-    // TODO: DCGM, EBS, and ServiceMonitor have Dynamic metric collection. During each iteration we have to
-    // check if these optionals have a set value. lets improve how we handle this
-
-    // Create a ServiceMonitor object to monitor Systemd services if any configs are valid
-    std::optional<ServiceMonitor> serviceMetrics{};
-    std::optional<std::vector<std::regex> > serviceConfig{
-        parse_service_monitor_config_directory(ServiceMonitorConstants::ConfigPath)};
-    if (serviceConfig.has_value())
-    {
-        serviceMetrics.emplace(registry, serviceConfig.value(), max_monitored_services);
-    }
-    else
-    {
-        Logger()->info("Service Monitoring is disabled.");
-    }
-
-    std::optional<Perfspect> perfspectMetrics{};
-    auto instanceInfo = Perfspect::IsValidInstance();
-    if (instanceInfo.has_value())
-    {
-        perfspectMetrics.emplace(registry, instanceInfo.value());
-    }
-    else
-    {
-        Logger()->info("PerfSpect Monitoring is disabled.");
-    }
-
-    // Create an EBS collector object to monitor EBS devices if any configs are valid
-    std::optional<EBSCollector> ebsMetrics{};
-    std::optional<std::unordered_set<std::string> > ebsConfig{parse_ebs_config_directory(EBSConstants::ConfigPath)};
-    if (ebsConfig.has_value())
-    {
-        ebsMetrics.emplace(registry, ebsConfig.value());
-    }
-    else
-    {
-        Logger()->info("EBS Monitoring is disabled.");
-    }
-
-    if (gpuDCGM.has_value())
-    {
-        std::string serviceStatus = atlasagent::is_service_running(DCGMConstants::ServiceName) ? "ON" : "OFF";
-        Logger()->info(
-            "DCGMI binary present. Agent will collect DCGM metrics if service is ON. DCGM service state: {}.",
-            serviceStatus);
-    }
-    else
-    {
-        Logger()->info("DCGMI binary not present. Agent will not collect DCGM metrics.");
-    }
-
-    auto gpuAMD = atlasagent::GpuMetricsAMD::Create(registry);
-    if (gpuAMD.has_value())
-    {
-        Logger()->info("AMD GPU(s) detected. Agent will collect AMD SMI metrics.");
-    }
-    else
-    {
-        Logger()->info("No AMD GPUs detected. Agent will not collect AMD SMI metrics.");
-    }
-
-    // initial polling delay, to prevent publishing too close to a minute boundary
-    auto delay = initial_polling_delay();
-    Logger()->info("Initial polling delay is {}s", delay);
-    if (delay > 0)
-    {
-        runner.wait_for(seconds(delay));
-    }
-
-    // the first call to this gather function takes ~100ms, so it must be
-    // done before we start calculating times to wait for peak metrics
-    gather_slow_system_metrics(&proc, &disk, &ethtool, &ntp, &pressureStall, &aws);
-    Logger()->info("Published slow system metrics (first iteration)");
-
-    auto now = system_clock::now();
-    auto next_run = now;
-    auto next_sixty_second_run = now + seconds(60);
-    auto next_five_second_run = now + seconds(5);
-    std::chrono::nanoseconds time_to_sleep;
-
-    do
-    {
-        auto start = system_clock::now();
-        bool fiveSecondMetricsEnabled = (start >= next_five_second_run);
-        bool sixtySecondMetricsEnabled = (start >= next_sixty_second_run);
-
-        // Gather one second metrics
-        // Proc has been modified to optionally gather 5 second and 60 second metrics during this call
-        // This prevents having to read proc/stat multiple times if both 5 and 60 second metrics are enabled
-        gather_peak_system_metrics(&proc, fiveSecondMetricsEnabled, sixtySecondMetricsEnabled);
-        gather_scaling_metrics(&cpufreq);
-
-        // If it's time to gather the 5 second metrics
-        if (fiveSecondMetricsEnabled == true)
-        {
-            Logger()->debug("Gathering 5 second metrics");
-            if (perfspectMetrics.has_value())
-            {
-                perfspectMetrics->GatherMetrics();
-            }
-            next_five_second_run += seconds(5);
-        }
-
-        // If it's time to gather the 60 second metrics
-        if (sixtySecondMetricsEnabled == true)
-        {
-            Logger()->debug("Gathering 60 second metrics");
-            gather_slow_system_metrics(&proc, &disk, &ethtool, &ntp, &pressureStall, &aws);
-            perf_metrics.collect();
-            if (gpu)
-            {
-                gpu->gpu_metrics();
-            }
-
-            if (gpuAMD.has_value())
-            {
-                gpuAMD->GPUMetrics();
-            }
-
-            if (gpuDCGM.has_value() && atlasagent::is_service_running(DCGMConstants::ServiceName))
-            {
-                if (gpuDCGM.value().gather_metrics() == false)
-                {
-                    Logger()->error("Failed to gather DCGM metrics");
-                }
-            }
-
-            if (ebsMetrics.has_value() && ebsMetrics.value().gather_metrics() == false)
-            {
-                Logger()->error("Failed to gather EBS metrics");
-            }
-
-            if (serviceMetrics.has_value() && serviceMetrics.value().gather_metrics() == false)
-            {
-                Logger()->error("Failed to gather Service metrics");
-            }
-
-            auto elapsed = duration_cast<milliseconds>(system_clock::now() - start);
-            Logger()->debug("Published system metrics (delay={})", elapsed);
-            next_sixty_second_run += seconds(60);
-        }
-
-        next_run += seconds(1);
-        time_to_sleep = next_run - system_clock::now();
-    } while (runner.wait_for(time_to_sleep));
-}
-#endif
 
 struct agent_options
 {
